requirements.txt
markdown==3.5.1
PyYAML==6.0.1
Jinja2==3.1.2 